+++ modules/contracts/src.ts/deployments.ts
@@ -6,6 +6,25 @@
 export const deployments = {} as any;
 
 ////////////////////////////////////////
+// 1 - mainnet
+import * as mainnetChannelFactory from "../deployments/mainnet/ChannelFactory.json";
+import * as mainnetChannelMastercopy from "../deployments/mainnet/ChannelMastercopy.json";
+import * as mainnetHashlockTransfer from "../deployments/mainnet/HashlockTransfer.json";
+import * as mainnetTestToken from "../deployments/mainnet/TestToken.json";
+import * as mainnetTransferRegistry from "../deployments/mainnet/TransferRegistry.json";
+import * as mainnetWithdraw from "../deployments/mainnet/Withdraw.json";
+const mainnetDeployment = {
+  ChannelFactory: mainnetChannelFactory,
+  ChannelMastercopy: mainnetChannelMastercopy,
+  HashlockTransfer: mainnetHashlockTransfer,
+  TestToken: mainnetTestToken,
+  TransferRegistry: mainnetTransferRegistry,
+  Withdraw: mainnetWithdraw,
+};
+deployments.mainnet = mainnetDeployment;
+deployments["1"] = mainnetDeployment;
+
+////////////////////////////////////////
 // 4 - rinkeby
 import * as rinkebyChannelFactory from "../deployments/rinkeby/ChannelFactory.json";
 import * as rinkebyChannelMastercopy from "../deployments/rinkeby/ChannelMastercopy.json";
@@ -61,6 +80,25 @@ const kovanDeployment = {
 };
 deployments.kovan = kovanDeployment;
 deployments["42"] = kovanDeployment;
+
+////////////////////////////////////////
+// 137 - matic
+import * as maticChannelFactory from "../deployments/matic/ChannelFactory.json";
+import * as maticChannelMastercopy from "../deployments/matic/ChannelMastercopy.json";
+import * as maticHashlockTransfer from "../deployments/matic/HashlockTransfer.json";
+import * as maticTestToken from "../deployments/matic/TestToken.json";
+import * as maticTransferRegistry from "../deployments/matic/TransferRegistry.json";
+import * as maticWithdraw from "../deployments/matic/Withdraw.json";
+const maticDeployment = {
+  ChannelFactory: maticChannelFactory,
+  ChannelMastercopy: maticChannelMastercopy,
+  HashlockTransfer: maticHashlockTransfer,
+  TestToken: maticTestToken,
+  TransferRegistry: maticTransferRegistry,
+  Withdraw: maticWithdraw,
+};
+deployments.matic = maticDeployment;
+deployments["137"] = maticDeployment;
 
 ////////////////////////////////////////
 // 80001 - mumbai