+++ modules/contracts/src.ts/tasks/index.ts
@@ -1,3 +1,4 @@
 import "./createChannel";
 import "./displayAccounts";
 import "./registerTransfer";
+import "./changeTransferRegistryOwner";